backend/requirements.txt
flask
flask-cors
python-dotenv
flask-bcrypt
pyjwt
pymongo
transformers
torch
